Fetch and cache Oncotree data on server startup
